better crypto
this adds the salt to the first part of the encoded buffer. this ought to fix a
problem where you could easily guess the input number by computing the md5sum
of small integers (say, all of those less than 1e6) and then comparing the
first 8 bytes of their md5sum to the first 8 bytes of the encrypted key.